+++ utils/energymon-power-poller.c
@@ -26,36 +26,35 @@ static const char* filename = NULL;
 static int summarize = 0;
 static uint64_t interval = 0;
 
-static const char* prog;
-static const char short_options[] = "c:f:i:sh";
+static const char short_options[] = "hc:f:i:s";
 static const struct option long_options[] = {
+  {"help",      no_argument,       NULL, 'h'},
   {"count",     required_argument, NULL, 'c'},
   {"file",      required_argument, NULL, 'f'},
   {"interval",  required_argument, NULL, 'i'},
   {"summarize", no_argument,       NULL, 's'},
-  {"help",      no_argument,       NULL, 'h'},
   {0, 0, 0, 0}
 };
 
 static void print_usage(int exit_code) {
   fprintf(exit_code ? stderr : stdout,
-          "Usage:\n"
-          "  %s [options]\n\n"
+          "Usage: energymon-power-poller [OPTION]...\n"
+          "Options:\n"
+          "  -h, --help               Print this message and exit\n"
           "  -c, --count=N            Stop after N reads\n"
           "  -f, --file=FILE          The output file\n"
-          "  -i, --interval=INTERVAL  The update interval in microseconds\n"
-          "  -s, --summarize          Print out a summary at completion\n"
-          "  -h, --help               Print this message and exit\n\n",
-          prog);
+          "  -i, --interval=US        The update interval in microseconds\n"
+          "  -s, --summarize          Print out a summary at completion\n");
   exit(exit_code);
 }
 
 static void parse_args(int argc, char** argv) {
   int c;
-  prog = argv[0];
-
   while ((c = getopt_long(argc, argv, short_options, long_options, NULL)) != -1) {
     switch (c) {
+      case 'h':
+        print_usage(0);
+        break;
       case 'c':
         count = 1;
         running = strtoull(optarg, NULL, 0);
@@ -65,30 +64,19 @@ static void parse_args(int argc, char** argv) {
         break;
       case 'i':
         interval = strtoull(optarg, NULL, 0);
+        if (interval == 0) {
+          fprintf(stderr, "Interval must be > 0\n");
+          print_usage(1);
+        }
         break;
       case 's':
         summarize = 1;
-        break;
-      case 'h':
-        print_usage(0);
         break;
       case '?':
       default:
         print_usage(1);
         break;
     }
-  }
-
-  // validate arguments
-  if (running == 0) {
-    // only happens if count was set to 0
-    fprintf(stderr, "Count must be > 0\n");
-    print_usage(1);
-  }
-  if (interval == 0) {
-    // happens if no interval was specified or set to 0
-    fprintf(stderr, "Interval must be set and be > 0\n");
-    print_usage(1);
   }
 }
 
@@ -132,22 +120,30 @@ int main(int argc, char** argv) {
   parse_args(argc, argv);
 
   // initialize the energy monitor
-  if (energymon_get_default(&em) || em.finit(&em)) {
+  if (energymon_get_default(&em)) {
+    perror("energymon_get_default");
+    return 1;
+  }
+  if (em.finit(&em)) {
+    perror("energymon:finit");
     return 1;
   }
 
-  // validate requested interval
+  // get the update interval
   min_interval = em.finterval(&em);
-  if (interval < min_interval) {
+  if (interval == 0) {
+    interval = min_interval;
+  } else if (interval < min_interval) {
     fprintf(stderr,
             "Requested interval is too short, minimum available: %"PRIu64"\n",
             min_interval);
+    em.ffinish(&em);
     return 1;
   }
 
   // open the output file
   if (filename != NULL) {
-    fout = fopen(filename, "w+");
+    fout = fopen(filename, "w");
     if (fout == NULL) {
       perror(filename);
       em.ffinish(&em);
@@ -166,7 +162,12 @@ int main(int argc, char** argv) {
     power = (energy - energy_last) / ((float) interval);
     if (fprintf(fout, "%f\n", power) < 0) {
       ret = 1;
-      perror(filename);
+      if (filename == NULL) {
+        // should never happen writing to stdout
+        perror("Writing");
+      } else {
+        perror(filename);
+      }
       break;
     }
     fflush(fout);
@@ -186,7 +187,9 @@ int main(int argc, char** argv) {
       pavg = pavg + ((power - pavg) / n);
     }
     pstd = pstd + (power - pavg) * (power - pavg_last);
-    energymon_sleep_us(interval, &IGNORE_INTERRUPT);
+    if (running) {
+      energymon_sleep_us(interval, &IGNORE_INTERRUPT);
+    }
   }
 
   if (summarize) {
@@ -201,11 +204,9 @@ int main(int argc, char** argv) {
   // cleanup
   if (filename != NULL && fclose(fout)) {
     perror(filename);
-    ret = 1;
   }
   if (em.ffinish(&em)) {
     perror("em.ffinish");
-    ret = 1;
   }
 
   return ret;