+++ RELEASES.md
@@ -6,6 +6,10 @@
  * Only export symbols in shared object libraries that are declared in public headers (prevents symbol name clashes)
  * Use GNU standard installation directories (multiarch support)
  * Increased minimum CMake version from 2.8 to 2.8.5 to support GNUInstallDirs
+ * Added man pages for utility applications; bug fixes and improvements
+  * energymon-cmd-profile: Accept commands without quotations; add help option
+  * energymon-file-provider: Add help, count, interval, and no-rewind options
+  * energymon-idle-power: Add help option
 
 ## v0.2.1 - 2017-10-01